+++ blosc/blosc.h
@@ -14,11 +14,11 @@
 /* Version numbers */
 #define BLOSC_VERSION_MAJOR    1    /* for major interface/format changes  */
 #define BLOSC_VERSION_MINOR    3    /* for minor interface/format changes  */
-#define BLOSC_VERSION_RELEASE  0    /* for tweaks, bug-fixes, or development */
+#define BLOSC_VERSION_RELEASE  1    /* for tweaks, bug-fixes, or development */
 
-#define BLOSC_VERSION_STRING   "1.3.0"  /* string version.  Sync with above! */
+#define BLOSC_VERSION_STRING   "1.3.1.dev"  /* string version.  Sync with above! */
 #define BLOSC_VERSION_REVISION "$Rev$"   /* revision version */
-#define BLOSC_VERSION_DATE     "$Date:: 2014-01-11 #$"    /* date version */
+#define BLOSC_VERSION_DATE     "$Date:: 2014-01-13 #$"    /* date version */
 
 #define BLOSCLZ_VERSION_STRING "1.0.1"   /* the internal compressor version */