+++ RELEASE_NOTES.rst
@@ -1,5 +1,5 @@
 ===============================
- Release notes for Blosc 1.3.0
+ Release notes for Blosc 1.3.1
 ===============================
 
 :Author: Francesc Alted
@@ -7,7 +7,14 @@
 :URL: http://www.blosc.org
 
 
-Changes from 1.2.3 to 1.3.0
+Changes from 1.3.0 to 1.3.1
+===========================
+
+#XXX version-specific blurb XXX#
+
+
+
+Changes from 1.2.4 to 1.3.0
 ===========================
 
 A nice handful of compressors have been added to Blosc: